add more to

+++ app/Http/Controllers/Dashboard/ImageController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Models\ProductImage;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 
@@ -27,8 +28,16 @@ class ImageController extends Controller
 
     public function cancel(Request $request) {
         $temp_path = public_path("uploads/temp/$request->image");
-        if(File::exists($temp_path)) {
-            File::delete($temp_path);
+        $product_path = public_path("uploads/product/$request->image");
+        if(File::exists($temp_path) || File::exists($product_path)) {
+             if(File::exists($temp_path)){
+                File::delete($temp_path);
+            }elseif(File::exists($product_path)){
+                //delete image from db
+                ProductImage::where('image',$request->image)->delete();
+                File::delete($product_path);
+               
+            }
             return response([
                 'status' => 200,
                 'message' => 'Image deleted successfully'
